+++ aa-display/src/main/java/io/github/nitsuya/aa/display/xposed/CoreManagerService.kt
@@ -12,6 +12,7 @@ import io.github.nitsuya.aa.display.ui.aa.split.SplitDisplayController
 import io.github.nitsuya.aa.display.ui.aa.split.SplitPane
 import io.github.nitsuya.aa.display.ui.window.DisplaySessionPolicy
 import io.github.nitsuya.aa.display.xposed.hook.PanePresentationGuard
+import io.github.nitsuya.aa.display.xposed.hook.VdImeDisplayPin
 import io.github.nitsuya.aa.display.xposed.util.Instances
 import io.github.nitsuya.aa.display.xposed.util.log
 import io.github.nitsuya.aa.display.xposed.util.logDebug
@@ -134,6 +135,7 @@ class CoreManagerService private constructor() : ICoreManager.Stub() {
             }
             Instances.init(systemContext)
             PanePresentationGuard.ensureHooked()
+            VdImeDisplayPin.ensureHooked()
         }
 
         fun isAaVirtualDisplay(displayId: Int): Boolean {